Leverage bulk delete api in bifrost
With bulk delete support in bifrost, the cleanup gen_fsm can now send arrays of
authz_ids to be deleted instead of one request per authz_id.  This endpoint
returns the count of what was deleted, which we ignore.  This cleanup action
does not require that the deletion be successful, only that the request is
made.

Any server or auth error from bifrost will result in the gen_fsm crashing.